+++ examples/get_moderation.rs
@@ -34,7 +34,7 @@ async fn run() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>>
 
     let broadcaster_id = token.user_id.as_str();
 
-    let client = HelixClient::with_client(surf::Client::new());
+    let client: HelixClient<surf::Client> = HelixClient::new();
 
     println!("====Moderators====");
     println!(