+++ examples/modify_channel.rs
@@ -35,7 +35,7 @@ async fn run() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>>
         .user_id
         .unwrap();
 
-    let client = HelixClient::with_client(surf::Client::new());
+    let client: HelixClient<surf::Client> = HelixClient::new();
 
     let req = twitch_api2::helix::channels::ModifyChannelInformationRequest::builder()
         .broadcaster_id(&*broadcaster_id)